requirements.txt
appdirs==1.4.3
cookies==2.2.1
dj-database-url==0.4.2
Django==1.10.6
django-cors-headers==1.3.1
djangorestframework==3.6.2
funcsigs==1.0.2
gunicorn==19.7.1
mock==2.0.0
packaging==16.8
pbr==2.0.0
psycopg2==2.7.1
pyparsing==2.2.0
requests==2.13.0
responses==0.5.1
six==1.10.0
slate==0.3
